+++ Code/config.hpp
@@ -40,7 +40,7 @@ typedef unsigned long ulong;
 typedef std::vector<uint> vec_uint;
 typedef std::vector<ulong> vec_ulong;
 typedef std::string jsonStr;
-typedef sf::Vector2i Vertex;
+typedef sf::Vector2f Vertex;
 typedef sf::Vector2i Vec2i;
 typedef sf::Vector2f Vec2f;
 typedef sf::IntRect AABB;
